Create initial assignment instructions (#1)
I went ahead and created initial assignment instructions for a basic git exercise that covers much of the content of lecture #2. Pull requests are skipped over since we didn't cover them yet, but we can change that if needed.

Further, this basic assignment does not include much handholding for the actual git commands but does walk folks through the workflow. We can tune up the difficulty if we feel it is necessary.